fix: 修改xml

+++ src/main/java/com/example/vivo_backend/service/UserService.java
@@ -3,7 +3,7 @@ package com.example.vivo_backend.service;
 import com.example.vivo_backend.vo.user.UserVO;
 
 public interface UserService {
-    Long login(UserVO loginForm);
+    int login(UserVO loginForm);
     void register(UserVO registerForm);
     Boolean existsTheUsername(String username);
 }
